generado de tabla de datos

+++ PFGigenaMauro/js/index.js
@@ -58,9 +58,17 @@ const products = [
   },
 ];
 
+const orderSummary = {
+  subtotal: 0,
+  shippingCost: 500,
+  total: 0,
+  discount: 0,
+}
+
 // Obtener productos desde localStorage al cargar la página
 const storedProducts = JSON.parse(localStorage.getItem("products")) || [];
 const storedCoupons = JSON.parse(localStorage.getItem("coupons")) || [];
+
 let storedCart = JSON.parse(localStorage.getItem("cart")) || {};
 
 // Verificar si el carrito está vacío y, si es así, inicializarlo con un objeto vacío
@@ -71,6 +79,10 @@ if (Object.keys(storedCart).length === 0) {
 // Guardar datos en localStorage al cargar la página (opcional, solo si el carrito estaba vacío inicialmente)
 if (Object.keys(storedCart).length === 0) {
   localStorage.setItem("cart", JSON.stringify(storedCart));
+}
+
+if (!localStorage.getItem("orderSummary")) {
+  localStorage.setItem("orderSummary", JSON.stringify(orderSummary));
 }
 
 // iteracion de los objetos en la pagina
@@ -133,22 +145,36 @@ const addToCartProduct = async (event) => {
   const product = storedProducts.find((item) => item.id == productId);
 
   if (product) {
-    if (storedCart[productId]) {
-      storedCart[productId]++; // Si el producto ya está en el carrito, incrementa la cantidad
-    } else {
-      storedCart[productId] = 1; // Si es la primera vez que se agrega, establece la cantidad en 1
-    }
-
-    // Guardar el carrito actualizado en localStorage
-    localStorage.setItem("cart", JSON.stringify(storedCart));
-
-    console.log("Carrito actualizado:", storedCart);
-
-    await Swal.fire({
-      icon: 'success',
-      title: 'Producto agregado al carrito',
-      text: `${product.name} se ha agregado al carrito.`,
+    // Mostrar la alerta de confirmación
+    const result = await Swal.fire({
+      title: '¿Agregar al carrito?',
+      text: `¿Deseas agregar ${product.name} al carrito?`,
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonText: 'ACEPTAR',
+      cancelButtonText: 'CANCELAR',
+      confirmButtonColor: "green",
+      cancelButtonColor: 'red'
     });
+
+    if (result.isConfirmed) {
+      if (storedCart[productId]) {
+        storedCart[productId]++; // Si el producto ya está en el carrito, incrementa la cantidad
+      } else {
+        storedCart[productId] = 1; // Si es la primera vez que se agrega, establece la cantidad en 1
+      }
+
+      // Guardar el carrito actualizado en localStorage
+      localStorage.setItem("cart", JSON.stringify(storedCart));
+
+      console.log("Carrito actualizado:", storedCart);
+
+      await Swal.fire({
+        icon: 'success',
+        title: 'Producto agregado al carrito',
+        text: `${product.name} se ha agregado al carrito.`,
+      });
+    }
   } else {
     await Swal.fire({
       icon: 'error',
